Added cursor bug and rate limit detection for Inventory History
Previous attempt to fix the cursor bug did not work and has been removed.  I'm not convinced a good fix exists.  It's a very hard bug to duplicate and even do any testing on.  I've documented it as much as I can and added detection to let the user figure it out for themselves.